refactor(ivy): split up ngcc parser classes to make the code clearer
